requirements.txt
google-auth-oauthlib
google-api-python-client
matplotlib
pandas
python-dotenv
pytest
